refactor: Move Boolean hypercube to separate module
